Added compression option for backup command
This commit introduces a new feature to the backup command in the CLI tool,
allowing users to compress their backups for more efficient storage management.
The --compress flag can be used to enable this option, which compresses
the backup directory into a tar.gz file and places it in the specified output directory.

Signed-off-by: Anton Kremenetsky <anton.kremenetsky@gmail.com>