Rotation and disk space guard features
Rotation
For the periodic backup, you can set the number of backups to keep(rotation number).
The default value is 5. Use the `--rotate` option to set the number of backups to keep:

Disk space gurad
Backups can take a lot of disk space and it can be a reason to crash the whole system
if a disk will be full. To prevent such a situation you can set a threshold for disk
space that should be free. If during a backup process this threshold will be reached,
the backup process will stop. Use the `--min-free-space` option to set the threshold
in GB.

Also the `offset` option was added to specify backup start time more convenient.

Signed-off-by: Anton Kremenetsky <anton.kremenetsky@gmail.com>