Added new path from env variable dependency
Similar to the `path` dependency but the actual path to file or
directory is specified via an environment variable:

```yaml
  deps:
    - dst: /opt/genesis_devtools
      path:
        env: LOCAL_AGENT_PATH
```

Also an additional parameter `optional` may be specified:

```
  deps:
    - dst: /opt/genesis_devtools
      optional: true
      path:
        env: LOCAL_AGENT_PATH
```

The optional dependency will be omitted if it absent. The main
motivation for this option is for local development when we would
like to put some local file, directory but for production build we will
take a package with specific version.

Signed-off-by: Anton Kremenetsky <anton.kremenetsky@gmail.com>